encode UTF-8 on the fly
We currently buffer up all of the UTF-32 encoded string, and then
UTF-8 encode the whole string in one go. This is simple, but
inefficient for two reasons:

1) We re-allocate the output-buffer each round of the loop
2) We need both string representations in memory at the same time

So let's fix this by allocating 16 bytes, increasing this by 150%
each time we need more, and re-encode to UTF-8 on the fly. This gives
us O(log n) allocations instaf of O(n) as we had before. The starting
point of 16 serves two purposes:

a) We avoid using proportionally many allocations for short strings
b) 150% of 16 is 24, which means we have 8 new characters of storage
   after the first re-allocation

The significance of b) is because the maximum number of new storage
bytes we could end up needing is 6: a high surrogate, followed by a
large BMP-character that is not a low surrogate. These will encode
to 3 bytes each, leaving us with needing 6 bytes, plus one for
termination.