free allocations on parse-errors
If we for instance failed to parse due to being out of memory, it's
not so nice to hog memory that we're not even going to reference
any more.

So let's release all memory in case of error, so the application
has a better chance of getting to do something useful.