fixup typo in UTF-8 encoding
In dbb3be8 ("parse and UTF-16 decode string, then UTF-8 encode"),
the loop-variable i was renamed to j, without updating the usage
in the encoding. Whoops.

Unfortunately, it seems the tests just blindly accepted broken
data. This time, the first part of the data has been verified to be
correct using other encoders as reference. The others are arguably
still wrong; the UTF-8 spec seems pretty strict about not allowing
to encode surrogate pairs. Patch for that coming up.